Add shader entry points to default pipeline names. (#186)
Currently the default pipeline name consists of the hashes of shaders in the pipeline. Add entry points to the name to possibly make it more readable.

The new behavior is enabled by default in the dumped event traces.
Overlay still uses shorter forms and can be switched to the new behavior in the settings tab.